Added predictive-latex-map and predictive-texinfo-map keymaps.
latex- and texinfo-specific predictive-mode bindings are now defined in those
keymaps, instead of in predictive-map. This fixes a bug that caused the
supposedly latex- and texinfo-specific bindings to be enabled in all
predictive-mode buffers, as the make-local-variable used previously doesn't
have the desired effect for keymaps.

Also defined predictive-latex-mode and predictive-texinfo-mode variables to
enable these new minor-mode keymaps. We don't use define-minor-mode here, as
we explicitly *don't* want to define an interactive toggling command.