Use with-local-exit in predictive-ispell-word-p, so C-g can escape hang.
Sometimes ispell-accept-output seems to work when called from an idle timer,
sometimes it hangs. Using predictive-ispell-word-p as a
predictive-auto-add-filter and enabling predictive-use-auto-learn-cache will
hit this problem. Since inhibit-quit is set when running timer functions, this
is particularly bad because even C-g won't save you! Protecting the call to
ispell-accept-output with with-local-exit at least allows C-g to work again.